Add signature for aisle config export
