+++ bindings/src/lib.rs
@@ -1,158 +1,15 @@
-use std::collections::HashMap;
+
 use cooklang::analysis::{parse_events, RecipeContent};
-use cooklang::model::{Item as ModelItem};
 use cooklang::parser::{PullParser};
-use cooklang::quantity::{
-    Quantity as ModelQuantity, ScalableValue as ModelScalableValue, Value as ModelValue,
-};
 use cooklang::Converter;
 use cooklang::Extensions;
 
-#[derive(uniffi::Record, Debug)]
-pub struct CooklangRecipe {
-    metadata: HashMap<String, String>,
-    steps: Vec<Step>,
-    ingredients: Vec<Item>,
-    cookware: Vec<Item>,
-}
+pub mod aisle;
+pub mod model;
 
-#[derive(uniffi::Record, Debug)]
-struct Step {
-    items: Vec<Item>,
-}
+use model::*;
+use aisle::*;
 
-#[derive(uniffi::Enum, Debug, Clone, PartialEq)]
-enum Item {
-    Text {
-        value: String,
-    },
-    Ingredient {
-        name: String,
-        amount: Option<Amount>,
-    },
-    Cookware {
-        name: String,
-        amount: Option<Amount>,
-    },
-    Timer {
-        name: Option<String>,
-        amount: Option<Amount>,
-    },
-}
-
-#[derive(uniffi::Record, Debug, Clone, PartialEq)]
-struct Amount {
-    quantity: Value,
-    units: Option<String>,
-}
-
-#[derive(uniffi::Enum, Debug, Clone, PartialEq)]
-enum Value {
-    Number { value: f64 },
-    Range { start: f64, end: f64 },
-    Text { value: String },
-}
-
-type CooklangMetadata = HashMap<String, String>;
-
-trait Amountable {
-    fn extract_amount(&self) -> Amount;
-}
-
-impl Amountable for ModelQuantity<ModelScalableValue> {
-    fn extract_amount(&self) -> Amount {
-        let quantity = extract_quantity(&self.value);
-
-        let units = if let Some(u) = &self.unit() {
-            Some(u.to_string())
-        } else {
-            None
-        };
-
-        Amount { quantity, units }
-    }
-}
-
-impl Amountable for ModelScalableValue {
-    fn extract_amount(&self) -> Amount {
-        let quantity = extract_quantity(&self);
-
-        Amount {
-            quantity,
-            units: None,
-        }
-    }
-}
-
-fn extract_quantity(value: &ModelScalableValue) -> Value {
-    match value {
-        ModelScalableValue::Fixed { value } => extract_value(value),
-        ModelScalableValue::Linear { value } => extract_value(value),
-        ModelScalableValue::ByServings { values } => extract_value(values.first().unwrap()),
-    }
-}
-
-fn extract_value(value: &ModelValue) -> Value {
-    match value {
-        ModelValue::Number { value } => Value::Number { value: *value },
-        ModelValue::Range { value } => Value::Range {
-            start: *value.start(),
-            end: *value.end(),
-        },
-        ModelValue::Text { value } => Value::Text {
-            value: value.to_string(),
-        },
-    }
-}
-
-
-fn into_item(item: ModelItem, recipe: &RecipeContent) -> Item {
-    match item {
-        ModelItem::Text { value } => Item::Text { value },
-        ModelItem::ItemIngredient { index } => {
-            let ingredient = &recipe.ingredients[index];
-
-            Item::Ingredient {
-                name: ingredient.name.clone(),
-                amount: if let Some(q) = &ingredient.quantity {
-                    Some(q.extract_amount())
-                } else {
-                    None
-                },
-            }
-        },
-
-        ModelItem::ItemCookware { index } => {
-            let cookware = &recipe.cookware[index];
-            Item::Cookware {
-                name: cookware.name.clone(),
-                amount: if let Some(q) = &cookware.quantity {
-                    Some(q.extract_amount())
-                } else {
-                    None
-                },
-            }
-        },
-
-        ModelItem::ItemTimer { index } => {
-            let timer = &recipe.timers[index];
-
-            Item::Timer {
-                name: timer.name.clone(),
-                amount: if let Some(q) = &timer.quantity {
-                    Some(q.extract_amount())
-                } else {
-                    None
-                },
-            }
-        },
-
-        // returning an empty block of text as it's not supported by the spec
-        ModelItem::InlineQuantity { .. } => Item::Text {
-            value: "".to_string(),
-        },
-    }
-}
 
 fn simplify_recipe_data(recipe: &RecipeContent) -> CooklangRecipe {
     let mut metadata = CooklangMetadata::new();
@@ -231,8 +88,13 @@ pub fn parse_metadata(input: String) -> CooklangMetadata {
     metadata
 }
 
+#[uniffi::export]
+pub fn parse_aisle_config(input: String) -> AisleConf {
+
+}
+
 // combine_amounts(amounts: Vec<Amount>) -> Vec<Amount>;
-// parse_aisle_config(input: String) -> AisleConfig;
+
 
 uniffi::setup_scaffolding!();
